refactor(tests): split test_models into tests/models/ package; share host_insight fixture
tests/test_models.py had grown to 701 lines. Split it per-domain to
mirror the models/ package - tests/models/test_{primitives,asset,network,
attack,h1,finding}.py - each well under the 450-line bar. Behaviour
unchanged; the classes move verbatim.

Also promotes the per-test _good_insight helper (duplicated across
test_recon_insights.py and the new test_recon_host_store.py) into a
shared make_host_insight factory fixture in tests/fixtures/recon.py,
per the fixtures-are-the-source-of-truth rule - one builder instead of
two drifting copies. Consumers take the fixture and call it; the
hostname derives from target_apex so the suite stays single-knob.
