--- tests/test_models.py
@@ -1,701 +1,0 @@
-"""
-tests/test_models.py - unit tests for models.py
-"""
-
-from __future__ import annotations
-
-import pytest
-from pydantic import BaseModel, ValidationError
-
-from models import (
-    FQDN,
-    AttackGraph,
-    Contact,
-    ContactRole,
-    Email,
-    Endpoint,
-    HttpUrl,
-    IPAddress,
-    RawFinding,
-    Severity,
-    VerifiedVulnerability,
-)
-from models.attack import AttackForest, AttackTree
-from models.h1 import (
-    DisclosureReport,
-    Programme,
-    ScopeItem,
-    ScopeType,
-    SubmissionResult,
-    SubmissionStatus,
-)
-
-pytestmark = pytest.mark.unit
-
-
-class TestSeverityEnum:
-    def test_all_values_present(self):
-        levels = {s.value for s in Severity}
-        assert levels == {"informational", "low", "medium", "high", "critical"}
-
-    def test_severity_is_string_enum(self):
-        assert isinstance(Severity.HIGH, str)
-        assert Severity.HIGH == "high"
-
-
-class TestScopeItem:
-    def test_valid_url_scope_item(self, scope_item_url):
-        assert scope_item_url.asset_type == ScopeType.URL
-        assert scope_item_url.eligible_for_bounty is True
-
-    def test_valid_wildcard_scope_item(self, scope_item_wildcard):
-        assert scope_item_wildcard.asset_type == ScopeType.WILDCARD
-
-    def test_instruction_defaults_to_none(self, target_apex):
-        item = ScopeItem(
-            asset_identifier=f"https://{target_apex}",
-            asset_type=ScopeType.URL,
-        )
-        assert item.instruction is None
-
-    def test_invalid_asset_type_raises(self, target_apex):
-        with pytest.raises(ValidationError):
-            ScopeItem(
-                asset_identifier=f"https://{target_apex}",
-                asset_type="not_a_real_type",
-            )
-
-
-class TestProgramme:
-    def test_valid_programme(self, programme):
-        assert programme.handle == "test-programme"
-        assert Severity.HIGH in programme.bounty_table
-
-    def test_priority_score_defaults_to_zero(self, programme):
-        assert programme.priority_score == 0.0
-
-    def test_selected_at_is_set(self, programme):
-        assert programme.selected_at is not None
-
-    def test_serialise_roundtrip(self, programme):
-        data = programme.model_dump()
-        restored = Programme.model_validate(data)
-        assert restored.handle == programme.handle
-        assert restored.bounty_table == programme.bounty_table
-
-
-class TestEndpoint:
-    def test_valid_endpoint(self, endpoint):
-        assert endpoint.status_code == 200
-        assert "nginx" in endpoint.technologies
-
-    def test_optional_fields_default(self, target_apex):
-        ep = Endpoint(url=f"https://{target_apex}")
-        assert ep.status_code is None
-        assert ep.technologies == []
-        assert ep.parameters == []
-
-
-# FQDN is exercised through a throwaway pydantic model so the validator
-# fires the same way it does when carried on a real schema field. Stick to
-# pytest.raises(ValidationError) rather than the lower-level ValueError so
-# the test mirrors what schema callers will see.
-class _FQDNProbe(BaseModel):
-    """Thin probe model used to drive the FQDN validator in isolation."""
-
-    value: FQDN
-
-
-class TestFQDN:
-    def test_accepts_target_apex(self, target_url):
-        # urlparse hostname is the canonical way to derive a FQDN-shaped
-        # string from a URL fixture; using the fixture keeps test intent
-        # ("the in-scope target") readable at the call site.
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        apex = host.split(".", 1)[-1]  # "example.com" from "victim.example.com"
-        assert _FQDNProbe(value=apex).value == apex
-
-    def test_accepts_victim_subdomain(self, target_url):
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        assert _FQDNProbe(value=host).value == host
-
-    def test_accepts_single_label(self):
-        # ``localhost`` is the canonical single-label hostname; no URL fixture
-        # exposes one because the rest of the codebase doesn't reach for it.
-        assert _FQDNProbe(value="localhost").value == "localhost"
-
-    def test_accepts_numeric_labels(self):
-        # 10.0.0.1 looks IP-shaped but parses as a valid hostname per RFC 1123
-        # label rules (digits are allowed). The scope filter is the next layer
-        # that decides whether to accept it as an in-scope target.
-        assert _FQDNProbe(value="10.0.0.1").value == "10.0.0.1"
-
-    def test_lowercases_victim_host(self, target_url):
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        assert _FQDNProbe(value=host.upper()).value == host
-
-    def test_strips_whitespace_around_victim_host(self, target_url):
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        assert _FQDNProbe(value=f"  {host}  ").value == host
-
-    def test_rejects_malformed(self, target_url):
-        """Walks the malformed corpus, deriving each case from target_url
-        so test intent ("a deliberately broken version of the in-scope
-        target") is readable. Pytest parametrize literals cannot consume
-        fixtures, so a single dedicated method loops the corpus instead.
-        """
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        cases: list[tuple[str, str]] = [
-            ("", "empty"),
-            ("   ", "whitespace only"),
-            (f"https://{host}", "scheme present"),
-            (f"ftp://{host}", "non-http scheme"),
-            (f"{host}:8080", "port present"),
-            (f"{host}/path", "path present"),
-            (f"{host}/", "trailing slash"),
-            (f"-{host}", "leading hyphen"),
-            (f"{host}-", "trailing hyphen on label"),
-            (host.replace(".", "..", 1), "empty label"),
-            ("a" * 64 + f".{host}", "label > 63 chars"),
-            (".".join(["a"] * 200), "total > 253 chars"),
-            (host.replace(".", " .", 1), "space in label"),
-            (f"{host}\nextra", "newline injection"),
-        ]
-        for value, label in cases:
-            with pytest.raises(ValidationError, match=r".*"):
-                _FQDNProbe.model_validate({"value": value})
-            # ``label`` is unused at the assertion level but appears in the
-            # case tuple so a future debugger can identify which case failed.
-            del label
-
-    def test_rejects_non_string(self):
-        with pytest.raises(ValidationError):
-            _FQDNProbe.model_validate({"value": 42})
-
-
-class _HttpUrlProbe(BaseModel):
-    """Thin probe model used to drive the HttpUrl validator in isolation."""
-
-    value: HttpUrl
-
-
-class TestHttpUrl:
-    def test_accepts_target_url(self, target_url):
-        assert _HttpUrlProbe(value=target_url).value == target_url
-
-    def test_accepts_target_url_with_path(self, target_url):
-        url = f"{target_url}/api/users?id=1"
-        assert _HttpUrlProbe(value=url).value == url
-
-    def test_accepts_http_scheme(self, target_url):
-        url = target_url.replace("https://", "http://")
-        assert _HttpUrlProbe(value=url).value == url
-
-    def test_rejects_malformed(self, target_url):
-        """Walks the malformed corpus, deriving each case from target_url so
-        intent ("a deliberately broken URL based on the in-scope target") is
-        readable at the call site. Delegates to ``pydantic.HttpUrl`` for the
-        URL contract; the host component then runs through the
-        ``FQDN`` validator so RFC 1123 strictness holds inside URLs
-        too - ``-evil.example.com`` rejects bare, and
-        ``https://-evil.example.com`` rejects wrapped (the leading-hyphen
-        case below pins that defense-in-depth).
-        """
-        from urllib.parse import urlparse
-
-        host = urlparse(target_url).hostname or ""
-        cases: list[tuple[str, str]] = [
-            ("", "empty"),
-            ("   ", "whitespace only"),
-            (host, "no scheme - bare hostname"),
-            (f"ftp://{host}", "non-http scheme"),
-            ("javascript:alert(1)", "javascript scheme"),
-            ("file:///etc/passwd", "file scheme"),
-            ("https://", "scheme with no host"),
-            ("https:///path", "scheme + path with no host"),
-            (f"https://-{host}", "hostname inside URL fails RFC 1123"),
-        ]
-        for value, label in cases:
-            with pytest.raises(ValidationError):
-                _HttpUrlProbe.model_validate({"value": value})
-            del label
-
-    def test_preserves_path_and_query(self, target_url):
-        url = f"{target_url}/search?q=hello&page=2#top"
-        assert _HttpUrlProbe(value=url).value == url
-
-
-class _IPAddressProbe(BaseModel):
-    """Thin probe model used to drive the IPAddress validator in isolation."""
-
-    value: IPAddress
-
-
-class TestIPAddress:
-    @pytest.mark.parametrize(
-        ("raw", "expected"),
-        [
-            # IPv4 dotted-quad
-            ("8.8.8.8", "8.8.8.8"),
-            ("1.1.1.1", "1.1.1.1"),
-            ("192.168.0.1", "192.168.0.1"),
-            ("0.0.0.0", "0.0.0.0"),
-            ("255.255.255.255", "255.255.255.255"),
-            # IPv6 - canonical form returns the compressed shape
-            ("::1", "::1"),
-            ("0:0:0:0:0:0:0:1", "::1"),
-            ("2001:db8::1", "2001:db8::1"),
-            ("2001:0db8:0000:0000:0000:0000:0000:0001", "2001:db8::1"),
-            # Whitespace tolerated and stripped
-            ("  8.8.8.8  ", "8.8.8.8"),
-        ],
-    )
-    def test_valid_address_canonicalises(self, raw, expected):
-        assert _IPAddressProbe(value=raw).value == expected
-
-    @pytest.mark.parametrize(
-        "bad",
-        [
-            "",
-            "   ",
-            "example.com",
-            "not-an-ip",
-            "1.2.3.4.5",
-            "256.0.0.1",
-            "1.2.3",
-            "999.999.999.999",
-        ],
-    )
-    def test_rejects_malformed(self, bad):
-        with pytest.raises(ValidationError):
-            _IPAddressProbe(value=bad)
-
-    def test_rejects_cidr_notation(self):
-        # CIDR is a netblock, not an IP; explicit reject so the error
-        # message names the cause.
-        with pytest.raises(ValidationError, match="CIDR"):
-            _IPAddressProbe(value="1.2.3.0/24")
-
-    def test_rejects_ipv6_zone_identifier(self):
-        # Link-local zone IDs (``fe80::1%eth0``) shouldn't appear in
-        # recon JSON - they're scoped to the scanner's host, not stable.
-        with pytest.raises(ValidationError, match="zone identifier"):
-            _IPAddressProbe(value="fe80::1%eth0")
-
-    def test_rejects_non_string_input(self):
-        with pytest.raises(ValidationError):
-            _IPAddressProbe.model_validate({"value": ["1.2.3.4"]})
-
-    def test_serialisation_roundtrip(self):
-        original = _IPAddressProbe(value="2001:db8::1")
-        restored = _IPAddressProbe.model_validate_json(original.model_dump_json())
-        assert restored.value == "2001:db8::1"
-
-    def test_runtime_type_is_str(self):
-        # IPAddress is Annotated[str, ...] - consumers that do
-        # f"https://{ip}" / ip.startswith(...) / dict-key keep working.
-        probe = _IPAddressProbe(value="8.8.8.8")
-        assert isinstance(probe.value, str)
-        assert probe.value.startswith("8.")
-
-
-class _EmailProbe(BaseModel):
-    """Thin probe model used to drive the Email validator in isolation."""
-
-    value: Email
-
-
-class TestEmail:
-    @pytest.mark.parametrize(
-        "raw",
-        [
-            "abuse@example.com",
-            "first.last@sub.example.com",
-            "a@b.co",
-        ],
-    )
-    def test_accepts_canonical_shapes(self, raw):
-        assert _EmailProbe(value=raw).value == raw
-
-    def test_normalises_to_lowercase(self):
-        # email_validator returns the normalised form (both local-part
-        # and domain case-folded) so equality holds across input
-        # variants.
-        assert _EmailProbe(value="Abuse@Example.COM").value == "abuse@example.com"
-
-    @pytest.mark.parametrize(
-        "bad",
-        ["", "   ", "no-at-sign", "@no-local-part.com", "missing-domain@", "two@@example.com"],
-    )
-    def test_rejects_malformed(self, bad):
-        with pytest.raises(ValidationError):
-            _EmailProbe(value=bad)
-
-    def test_runtime_type_is_str(self):
-        # Email is Annotated[str, ...] so f"mailto:{email}" /
-        # email.split("@") keep working without an audit.
-        probe = _EmailProbe(value="abuse@example.com")
-        assert isinstance(probe.value, str)
-        assert "@" in probe.value
-
-
-class TestContact:
-    def test_role_only_minimum(self):
-        # vCard fields are all optional - a Contact with just a role is
-        # legal but useless. The parser's _build_contact drops these
-        # before they reach RdapRecord; the model itself accepts.
-        c = Contact(role=ContactRole.ABUSE)
-        assert c.email is None
-        assert c.name is None
-        assert c.phone is None
-
-    def test_full_contact(self):
-        c = Contact(
-            role=ContactRole.REGISTRANT,
-            email="abuse@example.com",
-            name="Example Corp",
-            phone="+1-555-0123",
-        )
-        assert c.role is ContactRole.REGISTRANT
-        assert c.email == "abuse@example.com"
-
-    def test_invalid_email_rejects_contact(self):
-        # The Email primitive's validator fires on construction; a
-        # Contact carrying a mis-shaped email rejects rather than
-        # silently swallowing the bad value.
-        with pytest.raises(ValidationError):
-            Contact(role=ContactRole.ABUSE, email="not-an-email")
-
-    def test_serialise_roundtrip(self):
-        original = Contact(role=ContactRole.NOC, email="noc@example.com")
-        restored = Contact.model_validate_json(original.model_dump_json())
-        assert restored.role is ContactRole.NOC
-        assert restored.email == "noc@example.com"
-
-
-class TestAttackGraph:
-    def test_valid_recon_result(self, recon_result):
-        assert len(recon_result.subdomains) == 2
-        assert len(recon_result.endpoints) == 1
-
-    def test_completed_at_is_set(self, recon_result):
-        assert recon_result.completed_at is not None
-
-    def test_serialise_roundtrip(self, recon_result):
-        json_str = recon_result.model_dump_json()
-        restored = AttackGraph.model_validate_json(json_str)
-        assert restored.programme.handle == recon_result.programme.handle
-        assert len(restored.endpoints) == len(recon_result.endpoints)
-
-
-class TestIpAsset:
-    def test_ip_only_minimum(self):
-        # An IpAsset with only an IP is a useful starting record - the
-        # asn / rdap / ptr fields populate as enrichment completes.
-        from models import IpAsset
-
-        asset = IpAsset(ip="8.8.8.8")
-        assert asset.ip == "8.8.8.8"
-        assert asset.asn is None
-        assert asset.rdap is None
-        assert asset.ptr == []
-
-    def test_composes_typed_records(self, target_apex):
-        # The nested ``AsnRecord`` / ``RdapRecord`` / PTR hostnames
-        # arrive as separate lookups; IpAsset is the join point.
-        from models import IpAsset
-        from models.network import AsnRecord
-
-        asset = IpAsset(
-            ip="8.8.8.8",
-            asn=AsnRecord(
-                ip="8.8.8.8",
-                asn=15169,
-                prefix="8.8.8.0/24",
-                country="US",
-                organisation="GOOGLE, US",
-            ),
-            ptr=[f"api.{target_apex}"],
-        )
-        assert asset.asn is not None
-        assert asset.asn.asn == 15169
-        assert asset.ptr == [f"api.{target_apex}"]
-
-    def test_serialise_roundtrip(self):
-        from models import IpAsset
-
-        original = IpAsset(ip="1.1.1.1", ptr=["one.one.one.one"])
-        restored = IpAsset.model_validate_json(original.model_dump_json())
-        assert restored.ip == "1.1.1.1"
-        assert restored.ptr == ["one.one.one.one"]
-
-    def test_invalid_ip_rejects(self):
-        from pydantic import ValidationError
-
-        from models import IpAsset
-
-        with pytest.raises(ValidationError):
-            IpAsset(ip="not an ip")
-
-
-class TestService:
-    def test_minimal_record(self):
-        # host + port + protocol is the floor; the -sV banner fields and
-        # the technology rows populate only when the scan recovered them.
-        from models import Service
-
-        svc = Service(host="8.8.8.8", port=443, protocol="tcp")
-        assert svc.host == "8.8.8.8"
-        assert svc.port == 443
-        assert svc.name is None
-        assert svc.product is None
-        assert svc.technologies == []
-
-    def test_composes_typed_technologies(self, target_apex):
-        # The OA-boundary translation coerces banner detail into typed
-        # ``Technology`` rows hanging off the Service node.
-        from models import Service
-        from models.technology import Technology, TechnologyCategory
-
-        svc = Service(
-            host=f"api.{target_apex}",
-            port=443,
-            protocol="tcp",
-            name="http",
-            product="nginx",
-            version="1.25.3",
-            technologies=[
-                Technology(
-                    name="nginx",
-                    categories=[TechnologyCategory.web_server],
-                    version="1.25.3",
-                )
-            ],
-        )
-        assert svc.product == "nginx"
-        assert svc.technologies[0].name == "nginx"
-
-    def test_serialise_roundtrip(self):
-        from models import Service
-
-        original = Service(host="1.1.1.1", port=53, protocol="udp", name="domain")
-        restored = Service.model_validate_json(original.model_dump_json())
-        assert restored.host == "1.1.1.1"
-        assert restored.port == 53
-        assert restored.name == "domain"
-
-    def test_rejects_out_of_range_port(self):
-        from models import Service
-
-        with pytest.raises(ValidationError):
-            Service(host="8.8.8.8", port=70000, protocol="tcp")
-
-    def test_rejects_malformed_host(self):
-        # The FQDN | IPAddress union rejects a URL-shaped host upstream of
-        # any graph emission, the same boundary FQDN enforces elsewhere.
-        from models import Service
-
-        with pytest.raises(ValidationError):
-            Service(host="https://evil.test/x", port=443, protocol="tcp")
-
-    def test_rejects_oversize_banner(self):
-        # The max_length cap on the tool-captured banner fields is the
-        # boundary defence against an outsize injection riding a malformed
-        # service banner across the OA -> PT handoff.
-        from models import Service
-
-        with pytest.raises(ValidationError):
-            Service(host="8.8.8.8", port=443, protocol="tcp", product="x" * 129)
-
-
-class TestTLSCertificate:
-    def test_minimal_record(self, target_apex):
-        # host is the only floor; the grabbed cert detail populates as
-        # the tls-grab / testssl pass recovers it.
-        from models import TLSCertificate
-
-        cert = TLSCertificate(host=f"www.{target_apex}")
-        assert cert.host == f"www.{target_apex}"
-        assert cert.subject_common_name is None
-        assert cert.fingerprint_sha256 is None
-        assert cert.subject_alt_names == []
-
-    def test_captures_wildcard_san(self, target_apex):
-        # The whole reason SANs are list[str] not list[FQDN]: a wildcard
-        # SAN must survive into the cert asset, where list[FQDN] would
-        # reject it. iPAddress SANs land the same way.
-        from models import TLSCertificate
-
-        cert = TLSCertificate(
-            host=f"www.{target_apex}",
-            subject_common_name=f"www.{target_apex}",
-            issuer="Let's Encrypt",
-            subject_alt_names=[f"*.{target_apex}", target_apex, "203.0.113.7"],
-        )
-        assert f"*.{target_apex}" in cert.subject_alt_names
-        assert "203.0.113.7" in cert.subject_alt_names
-
-    def test_serialise_roundtrip(self, target_apex):
-        from models import TLSCertificate
-
-        original = TLSCertificate(
-            host=f"www.{target_apex}",
-            fingerprint_sha256="ab" * 32,
-            subject_alt_names=[f"*.{target_apex}"],
-        )
-        restored = TLSCertificate.model_validate_json(original.model_dump_json())
-        assert restored.fingerprint_sha256 == "ab" * 32
-        assert restored.subject_alt_names == [f"*.{target_apex}"]
-
-    def test_rejects_malformed_host(self):
-        from models import TLSCertificate
-
-        with pytest.raises(ValidationError):
-            TLSCertificate(host="https://evil.test/x")
-
-    def test_rejects_oversize_issuer(self, target_apex):
-        # The attacker-controlled cert string fields carry boundary caps.
-        from models import TLSCertificate
-
-        with pytest.raises(ValidationError):
-            TLSCertificate(host=f"www.{target_apex}", issuer="x" * 256)
-
-
-class TestHostScore:
-    def test_minimal(self, target_apex):
-        from models import HostPriority, HostRole, HostScore
-
-        s = HostScore(hostname=f"api.{target_apex}", role=HostRole.API, priority=HostPriority.HIGH)
-        assert s.role == HostRole.API
-        assert s.priority == HostPriority.HIGH
-        assert s.annotated_at is not None
-
-    def test_serialise_roundtrip(self, target_apex):
-        from models import HostPriority, HostRole, HostScore
-
-        original = HostScore(
-            hostname=f"auth.{target_apex}", role=HostRole.AUTH, priority=HostPriority.MEDIUM
-        )
-        restored = HostScore.model_validate_json(original.model_dump_json())
-        assert restored.hostname == f"auth.{target_apex}"
-        assert restored.role == HostRole.AUTH
-
-    def test_rejects_malformed_host(self):
-        from models import HostPriority, HostRole, HostScore
-
-        with pytest.raises(ValidationError):
-            HostScore(hostname="https://x/y", role=HostRole.API, priority=HostPriority.LOW)
-
-
-class TestAttackForest:
-    def test_serialise_roundtrip(self, attack_forest):
-        restored = AttackForest.model_validate_json(attack_forest.model_dump_json())
-        assert restored.programme_handle == attack_forest.programme_handle
-        assert restored.trees[0].expected_ceiling == Severity.CRITICAL
-
-
-class TestAttackTree:
-    def test_accepts_vulnerability_class_probe(self, target_url):
-        # The fixture covers CVE-id probes; this variant exercises the
-        # vulnerability-class name shape (the second canonical probe form)
-        # with a recon-evidence list to confirm the model accepts it end to end.
-        item = AttackTree(
-            probe="reflected XSS",
-            target=f"{target_url}/?q=test",
-            expected_ceiling=Severity.MEDIUM,
-            rationale="parameterised endpoint reflects q into response without escaping",
-            recon_evidence=[f"{target_url} hosts a Vue 2 SPA"],
-        )
-        assert item.probe == "reflected XSS"
-        assert item.expected_ceiling == Severity.MEDIUM
-
-    def test_recon_evidence_strips_and_filters_empties(self, target_url):
-        # The recon_evidence field carries a Pydantic field_validator:
-        # whitespace is trimmed off every entry, and empties are
-        # dropped. Every constructor (direct call, model_validate,
-        # model_validate_json on a re-loaded plan) sees the same
-        # cleaned list, so the wrapper does not need its
-        # own defensive shaping and the persisted artefact never carries
-        # whitespace-only entries.
-        item = AttackTree(
-            probe="reflected XSS",
-            target=f"{target_url}/?q=test",
-            expected_ceiling=Severity.MEDIUM,
-            rationale="parameterised endpoint reflects q into response without escaping",
-            recon_evidence=["  signal one  ", "", "   ", "signal two"],
-        )
-        assert item.recon_evidence == ["signal one", "signal two"]
-
-
-class TestRawFinding:
-    def test_valid_finding(self, raw_finding_high):
-        assert raw_finding_high.vuln_class == "SQLi"
-        assert raw_finding_high.severity_hint == Severity.HIGH
-
-    def test_severity_defaults_to_medium(self, target_apex):
-        finding = RawFinding(
-            title="Test",
-            vuln_class="XSS",
-            target=f"https://{target_apex}",
-            evidence="payload reflected",
-            tool="nuclei",
-        )
-        assert finding.severity_hint == Severity.MEDIUM
-
-
-class TestVerifiedVulnerability:
-    def test_valid_verified_vuln(self, verified_vuln):
-        assert verified_vuln.in_scope is True
-        assert verified_vuln.cvss_score == 8.8
-        assert len(verified_vuln.steps_to_reproduce) == 3
-
-    def test_confirmed_at_is_set(self, verified_vuln):
-        assert verified_vuln.confirmed_at is not None
-
-    def test_serialise_roundtrip(self, verified_vuln):
-        json_str = verified_vuln.model_dump_json()
-        restored = VerifiedVulnerability.model_validate_json(json_str)
-        assert restored.cvss_vector == verified_vuln.cvss_vector
-
-
-class TestDisclosureReport:
-    def test_valid_report(self, disclosure_report):
-        assert disclosure_report.programme_handle == "test-programme"
-        assert disclosure_report.weakness_id == 89
-
-    def test_attachments_default_empty(self, disclosure_report):
-        assert disclosure_report.attachments == []
-
-    def test_serialise_roundtrip(self, disclosure_report):
-        json_str = disclosure_report.model_dump_json()
-        restored = DisclosureReport.model_validate_json(json_str)
-        assert restored.title == disclosure_report.title
-
-
-class TestSubmissionResult:
-    def test_default_status_is_pending(self):
-        result = SubmissionResult()
-        assert result.status == SubmissionStatus.PENDING
-        assert result.report_id is None
-
-    def test_successful_submission(self):
-        from datetime import UTC, datetime
-
-        result = SubmissionResult(
-            report_id="12345",
-            status=SubmissionStatus.SUBMITTED,
-            h1_url="https://hackerone.com/reports/12345",
-            submitted_at=datetime.now(UTC),
-        )
-        assert result.report_id == "12345"
-        assert result.submitted_at is not None
